Separate import window into own module
